refactor(filter): consolidate HTTP hook runners in http_utils
Move per-hook execution into shared runners with a unified
HeaderFilterOutcome for request and response headers. Gate TCP
filters and request conditions in execute_http_request so
http.rs only handles pipeline iteration and control flow.
